Remove phot. metal scaling methods add `chemistry`
Scaling the photospheric abundances with total metallicity Z is not principled because the models fix the protostellar abundances, and the photospheric abundances can diverge due to diffusive processes over the life of a star. Because this effect is time-dependent, it is correct only to scale the initial abundances with Z, not the photospheric ones.